backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
drf-extra-fields==3.7.0
django-cors-headers==4.3.1
django-colorfield==0.11.0
django-filter==23.3
djoser==2.1.0
gunicorn==20.1.0
isort==5.13.2
pep8==1.7.1